+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 
-## [1.1.2] - unreleased
+## [1.2.0] - unreleased
+
+### Added
+
+- Support for config files written in jsonnet ([#542](https://github.com/eclipse-csi/otterdog/pull/542))
 
 ## [1.1.1] - 21/11/2025